UpdateNonceManagers
